Remove bulk processing modules and related functionality
This commit deletes the `bulk_format`, `bulk`, `iiif_bulk_parser`, and `simple_text_parser` modules, along with their associated code and tests. This change simplifies the codebase by removing unused components, streamlining the overall structure, and enhancing maintainability. The removal of these modules is part of a broader effort to refactor and optimize the application.
