+++ src/bulk/mod.rs
@@ -1,0 +1,11 @@
+pub mod content_reader;
+pub mod output_path;
+pub mod parsers;
+pub mod processor;
+pub mod types;
+
+// Re-export the main public APIs
+pub use content_reader::{read_bulk_urls, read_urls_from_content_with_parsers};
+pub use output_path::generate_output_path_for_item;
+pub use processor::process_bulk;
+pub use types::{BulkInputParser, BulkParser, BulkProcessedItem};
