+++ src/bulk/parsers/mod.rs
@@ -1,0 +1,2 @@
+pub mod iiif_manifest;
+pub mod simple_text;
